feat(fe): Add "Note by" in the client summary - Cleaned up code
